+++ src/app/bungie-api/bungie-api-utils.js
@@ -1,4 +1,4 @@
-let apiKey;
+export let apiKey;
 
 if ($DIM_FLAVOR === 'release' || $DIM_FLAVOR === 'beta') {
   apiKey = $DIM_WEB_API_KEY;