+++ docs/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Sorting characters by age should be correct for D2 on PC.
 * The infusion fuel finder now supports reverse lookups, so you can choose the best thing to infuse a particular item *into*.
+* Your character stats now update live when you change armor.
 
 # 4.33.1